Made css parser ignore [], found some bugs and added comments.
Made the css parser ignore [] and everything between it, meaning it
treats it as part of the selector string name. I also commented quite a
bit of code in the css parser part. Selectors with ',' are currently
still not working correcty.(read todo in cssparse.cpp) Also found
another bug which is described in cssparser.cpp and need to be
fixed.(bug in the ".ca.cb.cc {}" part.)