New url decoder, can now click links. Only partially working tho.
Can now click and load links but only partially works. If the link is
moved down by e.g. a bigger element on the same line, the boudning box
stays as the collision is immideatly checked(not after layout). Also it
always takes the last link because that's the last saved link text.